% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_trend_by_years.R
\name{overall_trend_by_years}
\alias{overall_trend_by_years}
\title{Create a line graph to show scoring trend from beginning.}
\usage{
overall_trend_by_years()
}
\description{
Create a line graph to show scoring trend from beginning.
}
