% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_by_team.R
\name{record_by_team}
\alias{record_by_team}
\title{Show how many games are played after enter particular season and show game results.}
\usage{
record_by_team(college, X)
}
\arguments{
\item{college}{character string}

\item{X}{numeric variable}
}
\description{
Show how many games are played after enter particular season and show game results.
}
