% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_by_season.R
\name{stats_by_season}
\alias{stats_by_season}
\title{Show the home team season stats when enter particular season.}
\usage{
stats_by_season(X)
}
\arguments{
\item{X}{numeric variable}
}
\description{
Show the home team season stats when enter particular season.
}
